Refactor Dutch researcher
