+++ packages/yoastseo/spec/morphology/dutch/getStemWordsWithTAndDEndingSpecs.js
@@ -1,4 +1,4 @@
-import { generateCorrectStemWithTAndDEnding } from "../../../src/languageProcessing/languages/nl/morphology/getStemWordsWithTAndDEnding";
+import { generateCorrectStemWithTAndDEnding } from "../../../src/languageProcessing/languages/nl/helpers/internal/getStemWordsWithTAndDEnding";
 import getMorphologyData from "../../specHelpers/getMorphologyData";
 
 const morphologyDataNL = getMorphologyData( "nl" ).nl;
