+++ packages/yoastseo/spec/morphology/dutch/stemTOrDFromEndOfWordSpec.js
@@ -1,4 +1,4 @@
-import { stemTOrDFromEndOfWord } from "../../../src/languageProcessing/languages/nl/morphology/stemTOrDFromEndOfWord";
+import { stemTOrDFromEndOfWord } from "../../../src/languageProcessing/languages/nl/helpers/internal/stemTOrDFromEndOfWord";
 import getMorphologyData from "../../specHelpers/getMorphologyData";
 
 const morphologyDataNL = getMorphologyData( "nl" ).nl;
