+++ packages/yoastseo/src/languageProcessing/languages/nl/config/internal/passiveVoiceAuxiliaries.js
@@ -2,13 +2,12 @@
  * Returns a list with auxiliaries for the Dutch passive voice assessment.
  * @returns {Array} The list with auxiliaries.
  */
-export default function() {
-	return [
-		"word",
-		"wordt",
-		"worden",
-		"werd",
-		"werden",
-		"wordend",
-	];
-}
+export default [
+	"word",
+	"wordt",
+	"worden",
+	"werd",
+	"werden",
+	"wordend",
+];
+
