+++ packages/yoastseo/src/languageProcessing/languages/nl/helpers/getSentenceParts.js
@@ -1,15 +1,14 @@
 import splitSentence from "../../../helpers/passiveVoice/periphrastic/getSentencePartsSplitOnStopwords";
 import arrayToRegex from "../../../helpers/regex/createRegexFromArray";
-import SentencePart from "../config/passiveVoice/SentencePart";
-import auxiliariesFactory from "../config/passiveVoice/auxiliaries.js";
-import stopwordsFactory from "../config/stopwords.js";
+import SentencePart from "../values/SentencePart";
+import auxiliaries from "../config/internal/passiveVoiceAuxiliaries.js";
+import stopwords from "../config/stopWords.js";
 
 const options = {
 	SentencePart: SentencePart,
-	locale: "nl_NL",
 	regexes: {
-		auxiliaryRegex: arrayToRegex( auxiliariesFactory().all ),
-		stopwordRegex: arrayToRegex( stopwordsFactory() ),
+		auxiliaryRegex: arrayToRegex( auxiliaries ),
+		stopwordRegex: arrayToRegex( stopwords ),
 	},
 };
 
@@ -17,9 +16,9 @@ const options = {
  * Gets the sentence parts from a sentence by determining sentence breakers.
  *
  * @param {string} sentence The sentence to split up in sentence parts.
- * @param {Object} options The language options for which to get the sentence parts.
+ *
  * @returns {Array} The array with all parts of a sentence that have an auxiliary.
  */
-const getSentenceParts = function( sentence) {
+export default function getSentenceParts( sentence ) {
 	return splitSentence( sentence, options );
-};
+}
