+++ packages/yoastseo/src/languageProcessing/languages/nl/helpers/internal/stemTOrDFromEndOfWord.js
@@ -1,7 +1,7 @@
-import { checkIfWordEndingIsOnExceptionList, checkIfWordIsOnVerbExceptionList } from "../morphoHelpers/exceptionListHelpers";
+import { checkIfWordEndingIsOnExceptionList, checkIfWordIsOnVerbExceptionList } from "../../../../helpers/morphology/exceptionListHelpers";
 import { detectAndStemRegularParticiple } from "./detectAndStemRegularParticiple";
 import { generateCorrectStemWithTAndDEnding } from "./getStemWordsWithTAndDEnding";
-import checkExceptionsWithFullForms from "../morphoHelpers/checkExceptionsWithFullForms";
+import checkExceptionsWithFullForms from "../../../../helpers/morphology/checkExceptionsWithFullForms";
 
 /**
  * If the word ending in -t/-d was not matched in any of the checks for whether -t/-d should be stemmed or not, other checks still need
